requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
gunicorn>=21.0.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
pydantic>=2.5.0
scikit-learn>=1.4.0
numpy>=1.26.0
google-generativeai>=0.4.0
python-dotenv>=1.0.0
httpx>=0.27.0
shapely>=2.0.0
python-multipart>=0.0.6
